# vectispire-angular/openapi.json
{
  "components" : {
    "schemas" : {
...
      "TeamRequest" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TeamSummary" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "memberCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "notified" : {
            "type" : "boolean"
          },
          "targetCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "TeamTargetAssignment" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserTargetAssignment" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "WebhookRequest" : {
        "properties" : {
          "url" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "AgentKeyAuth" : {
        "description" : "Pre-shared authentication key for remote scanning agents",
        "in" : "header",
        "name" : "X-Agent-Key",
        "type" : "apiKey"
      },
      "ApiKeyAuth" : {
        "description" : "Programmatic API Key for CI/CD and automation integrations",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "BearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "JWT Session Bearer token obtained via /api/v1/auth/login",
        "name" : "BearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "name" : "Asmolabs Security Team",
      "url" : "https://github.com/asmolabs/vectispire"
    },
    "description" : "Automated supply-chain security, SBOM analysis, vulnerability triage, attack surface inventory, and compliance governance platform.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Vectispire Control Plane REST API",
    "version" : "4.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/teams" : {
      "get" : {
        "operationId" : "list_2",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeamSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      },
      "post" : {
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamSummary"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "teams-controller" ]
      }
    },
    "/api/v1/teams/{id}" : {
      "delete" : {
        "operationId" : "remove_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "teams-controller" ]
      },
      "patch" : {
        "operationId" : "rename",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      }
    },
    "/api/v1/teams/{id}/members" : {
      "get" : {
        "operationId" : "members",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "format" : "int64",
                    "type" : "integer"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      },
      "put" : {
        "operationId" : "setMembers",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "format" : "int64",
                  "type" : "integer"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "format" : "int64",
                    "type" : "integer"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      }
    },
    "/api/v1/teams/{id}/targets" : {
      "get" : {
        "operationId" : "targets_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeamTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      },
      "put" : {
        "operationId" : "setTargets_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/TeamTargetAssignment"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TeamTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      }
    },
    "/api/v1/teams/{id}/webhook" : {
      "put" : {
        "operationId" : "setWebhook",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WebhookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "teams-controller" ]
      }
    },
...
    "/api/v1/users/{id}/targets" : {
      "get" : {
        "operationId" : "targets",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UserTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "users-controller" ]
      },
      "put" : {
        "operationId" : "setTargets",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/UserTargetAssignment"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/UserTargetAssignment"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "users-controller" ]
      }
    },
...
  },
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Current Vectispire Instance",
    "url" : "/"
  } ],
...
}